处理登陆错误信息

+++ wms-user/src/main/java/com/lcdt/cwms/user/service/CompanyService.java
@@ -14,6 +14,6 @@ public interface CompanyService {
 
 	 WmsCompany createCompany(CreateCompanyDto createCompanyDto) throws UserNotExistException;
 
-	 List<WmsCompanyUserRelation> userCompanys(Integer userId);
+	 List<WmsCompanyUserRelation> userCompanys(Long userId);
 
 }
